+++ Patches/EnemyHealthPatch.cs
@@ -8,7 +8,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch("Hurt")]
         [HarmonyPrefix]
         private static void Hurt_Prefix(ref int _damage, int ___healthCurrent) {
-            if (Settings.weakEnemies) {
+            if (Settings.instance.weakEnemies) {
                 _damage = ___healthCurrent;
                 // sanity check
                 if (___healthCurrent < 1)