+++ Patches/EnemyVisionPatch.cs
@@ -1,8 +1,4 @@
 ﻿using HarmonyLib;
-using RepoAdminMenu.Utils;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace RepoAdminMenu.Patches {
 
@@ -13,7 +9,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch("VisionTrigger")]
         [HarmonyPrefix]
         private static bool VisionTrigger_Prefix(int playerID, PlayerAvatar player, EnemyVision __instance) {
-            if(Settings.blindEnemies || PlayerUtil.isNoTarget(player)) {
+            if(Settings.instance.blindEnemies || Settings.isNoTarget(player)) {
                 __instance.VisionTriggered[playerID] = false;
                 __instance.VisionsTriggered[playerID] = 0;
                 return false;