+++ Patches/NoTargetPatches.cs
@@ -1,10 +1,4 @@
 ﻿using HarmonyLib;
-using RepoAdminMenu.Utils;
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Text;
-using UnityEngine.UIElements;
 
 namespace RepoAdminMenu.Patches {
 
@@ -14,7 +8,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(Enemy), "Update")]
         [HarmonyPrefix]
         private static void Enemy_Update_Prefix(ref PlayerAvatar ___TargetPlayerAvatar) {
-            if (___TargetPlayerAvatar != null && PlayerUtil.isNoTarget(___TargetPlayerAvatar)) {
+            if (___TargetPlayerAvatar != null && Settings.isNoTarget(___TargetPlayerAvatar)) {
                 ___TargetPlayerAvatar = null;
             }
         }
@@ -22,7 +16,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyAnimal), "Update")]
         [HarmonyPrefix]
         private static void EnemyAnimal_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -30,7 +24,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyBangDirector), "Update")]
         [HarmonyPrefix]
         private static void EnemyBangDirector_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -38,7 +32,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyBeamer), "Update")]
         [HarmonyPrefix]
         private static void EnemyBeamer_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -46,7 +40,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyBowtie), "Update")]
         [HarmonyPrefix]
         private static void EnemyBowtie_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -54,7 +48,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyCeilingEye), "Update")]
         [HarmonyPrefix]
         private static void EnemyCeilingEye_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -62,7 +56,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyDuck), "Update")]
         [HarmonyPrefix]
         private static void EnemyDuck_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -70,7 +64,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyDuck), "OnGrabbed")]
         [HarmonyPrefix]
         private static bool EnemyDuck_OnGrabbed_Prefix(EnemyDuck __instance) {
-            if (Settings.friendlyDuck) {
+            if (Settings.instance.friendlyDuck) {
                 __instance.currentState = EnemyDuck.State.Roam;
                 return false;
             }
@@ -80,7 +74,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyDuck), "UpdateState")]
         [HarmonyPrefix]
         private static void EnemyDuck_UpdateState_Prefix(ref EnemyDuck.State _state) {
-            if (Settings.friendlyDuck && _state == EnemyDuck.State.AttackStart) {
+            if (Settings.instance.friendlyDuck && _state == EnemyDuck.State.AttackStart) {
                 _state = EnemyDuck.State.Roam;
             }
         }
@@ -88,7 +82,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyFloater), "Update")]
         [HarmonyPrefix]
         private static void EnemyFloater_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -96,7 +90,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyGnomeDirector), "Update")]
         [HarmonyPrefix]
         private static void EnemyGnomeDirector_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -104,7 +98,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyHidden), "Update")]
         [HarmonyPrefix]
         private static void EnemyHidden_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -112,7 +106,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyRobe), "Update")]
         [HarmonyPrefix]
         private static void EnemyRobe_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -120,7 +114,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyRunner), "Update")]
         [HarmonyPrefix]
         private static void EnemyRunner_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -128,7 +122,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemySlowMouth), "Update")]
         [HarmonyPrefix]
         private static void EnemySlowMouth_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -136,7 +130,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemySlowWalker), "Update")]
         [HarmonyPrefix]
         private static void EnemySlowWalker_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -144,7 +138,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyThinMan), "Update")]
         [HarmonyPrefix]
         private static void EnemyThinMan_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }
@@ -152,7 +146,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyTumbler), "Update")]
         [HarmonyPrefix]
         private static void EnemyTumbler_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -160,7 +154,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyUpscream), "Update")]
         [HarmonyPrefix]
         private static void EnemyUpscream_Update_Prefix(ref PlayerAvatar ___targetPlayer) {
-            if (___targetPlayer != null && PlayerUtil.isNoTarget(___targetPlayer)) {
+            if (___targetPlayer != null && Settings.isNoTarget(___targetPlayer)) {
                 ___targetPlayer = null;
             }
         }
@@ -168,7 +162,7 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch(typeof(EnemyValuableThrower), "Update")]
         [HarmonyPrefix]
         private static void EnemyValuableThrower_Update_Prefix(ref PlayerAvatar ___playerTarget) {
-            if (___playerTarget != null && PlayerUtil.isNoTarget(___playerTarget)) {
+            if (___playerTarget != null && Settings.isNoTarget(___playerTarget)) {
                 ___playerTarget = null;
             }
         }