+++ Patches/PlayerControllerPatch.cs
@@ -1,9 +1,4 @@
 ﻿using HarmonyLib;
-using RepoAdminMenu.Utils;
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Text;
 
 namespace RepoAdminMenu.Patches {
 
@@ -14,11 +9,11 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPrefix]
         private static void Update_Prefix(PlayerController __instance, PlayerAvatar ___playerAvatarScript) {
             if (___playerAvatarScript != null){
-                if (PlayerUtil.isInfiniteStamina(___playerAvatarScript)) {
+                if (Settings.isInfiniteStamina(___playerAvatarScript)) {
                     __instance.EnergyCurrent = __instance.EnergyStart;
                 }
 
-                __instance.DebugNoTumble = PlayerUtil.isNoTumble(___playerAvatarScript);
+                __instance.DebugNoTumble = Settings.isNoTumble(___playerAvatarScript);
             }
         }
     }