+++ Patches/PlayerHealthPatch.cs
@@ -1,5 +1,4 @@
 ﻿using HarmonyLib;
-using RepoAdminMenu.Utils;
 using System;
 
 namespace RepoAdminMenu.Patches {
@@ -10,11 +9,11 @@ namespace RepoAdminMenu.Patches {
         [HarmonyPatch("Hurt")]
         [HarmonyPrefix]
         private static bool Hurt_Prefix(ref int damage, PlayerAvatar ___playerAvatar, int ___health) {
-            if (PlayerUtil.isGod(___playerAvatar)) {
+            if (Settings.isGod(___playerAvatar)) {
                 return false;
             }
 
-            if (PlayerUtil.isNoDeath(___playerAvatar)) {
+            if (Settings.isNoDeath(___playerAvatar)) {
                 damage = Math.Min(damage, ___health - 1);
                 if (damage == 0) {
                     return false;