+++ Utils/EnemyUtil.cs
@@ -1,13 +1,8 @@
-﻿using HarmonyLib;
-using RepoAdminMenu.Patches;
+﻿using RepoAdminMenu.Patches;
 using REPOLib.Modules;
-using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Text;
 using UnityEngine;
-using UnityEngine.UIElements;
 
 namespace RepoAdminMenu.Utils {
     internal class EnemyUtil {